requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.6
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.8
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-js-asset==1.2.2
future==0.18.2
html5lib==1.1
importlib-metadata==1.7.0
kombu==4.6.11
Pillow==8.3.1
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2020.1
# reportlab==3.5.44
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
vine==1.3.0
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.4
zipp==3.1.0
